requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
--index-url https://pkgs.safetycli.com/repository/pekim/pypi/simple/

alembic==1.16.5
    # via flask-migrate
better-profanity==0.7.0
    # via -r requirements.in
bleach==6.3.0
    # via -r requirements.in
blinker==1.9.0
    # via
    #   flask
    #   flask-mail
boto3==1.40.37
    # via -r requirements.in
botocore==1.40.37
    # via
    #   boto3
    #   s3transfer
cachelib==0.13.0
    # via flask-caching
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via flask
deprecated==1.3.1
    # via limits
flask==3.1.2
    # via
    #   -r requirements.in
    #   flask-caching
    #   flask-limiter
    #   flask-login
    #   flask-mail
    #   flask-migrate
    #   flask-sqlalchemy
    #   flask-wtf
flask-caching==2.3.1
    # via -r requirements.in
flask-limiter==4.1.1
    # via -r requirements.in
flask-login==0.6.3
    # via -r requirements.in
flask-mail==0.10.0
    # via -r requirements.in
flask-migrate==4.1.0
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via
    #   -r requirements.in
    #   flask-migrate
flask-wtf==1.2.2
    # via -r requirements.in
greenlet==3.2.4
    # via sqlalchemy
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.6
    # via flask
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
limits==5.8.0
    # via flask-limiter
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
ordered-set==4.1.0
    # via flask-limiter
packaging==25.0
    # via
    #   gunicorn
    #   limits
psycopg2-binary==2.9.10
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.1.1
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
s3transfer==0.14.0
    # via boto3
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.43
    # via
    #   alembic
    #   flask-sqlalchemy
typing-extensions==4.15.0
    # via
    #   alembic
    #   flask-limiter
    #   limits
    #   sqlalchemy
urllib3==2.5.0
    # via
    #   botocore
    #   requests
webencodings==0.5.1
    # via bleach
werkzeug==3.1.3
    # via
    #   flask
    #   flask-login
wrapt==2.1.2
    # via deprecated
wtforms==3.2.1
    # via flask-wtf
